+++ patterns--result/Patterns.Result/ProblemExtensions.cs
@@ -1,0 +1,10 @@
+namespace Patterns.Result;
+
+internal static class ProblemExtensions
+{
+    public static IProblem ToProblem(this Exception ex)
+        => new Problem(ex.GetType().Name, ex.Message, default, ProblemSeverity.Error);
+
+    public static List<IProblem> ToProblems(this Exception ex)
+        => [ex.ToProblem()];
+}